PostgreSQL_AutoML/requirements.txt
psycopg2
mljar-supervised